requirements.txt
# pip install -r requirements.txt

Django>=4.2,<5.0
python-dotenv>=1.0.1
google-generativeai>=0.3.2
